--- src/main/java/com/mycompany/proyecto4h/Proyecto4H.java
@@ -1,20 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- */
-package com.mycompany.proyecto4h;
-
-import vista.VistaArrastraSilaba;
-
-/**
- *
- * @author Alan
- */
-public class Proyecto4H {
-
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-        System.out.println("Hello GitHub");
-        
-        java.awt.EventQueue.invokeLater(() -> new VistaArrastraSilaba().setVisible(true));
-    }
-}
