+++ src/main/java/controlador/ControladorInvitado.java
@@ -1,0 +1,24 @@
+
+package controlador;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import vista.Invitado;
+
+/**
+ *
+ * @author ep712
+ */
+public class ControladorInvitado implements ActionListener{
+    
+    Invitado objInvitado;
+
+    public ControladorInvitado(Invitado objInvitado) {
+        this.objInvitado = objInvitado;
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        
+    }            
+}
