+++ src/main/java/controlador/ControladorRegistro.java
@@ -1,0 +1,20 @@
+package controlador;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import vista.Registro;
+
+public class ControladorRegistro implements ActionListener {
+
+    Registro objRegistro;
+
+    public ControladorRegistro(Registro objRegistro) {
+        this.objRegistro = objRegistro;
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+    }
+
+}
